+++ next.config.mjs
@@ -1,8 +1,8 @@
 import MDX from '@next/mdx'
+import withNextjsStaticProps from 'recma-nextjs-static-props'
 import withFrontmatter from 'remark-frontmatter'
 import withGitHubFlavoredMarkdown from 'remark-gfm'
 import withMdxFrontmatter from 'remark-mdx-frontmatter'
-import withNextjsStaticProps from 'recma-nextjs-static-props'
 
 const withMDX = MDX({
   extension: /\.mdx$/,
