+++ pages/concepts/planetscale-workflow.mdx
@@ -6,11 +6,7 @@ slug: 'concepts/planetscale-workflow'
 date: '2022-07-22'
 ---
 
-import ConceptLayout from '../../layouts/ConceptLayout'
-
-export default function Layout(props) {
-  return <ConceptLayout {...props} />
-}
+export { default } from '../../layouts/PostLayout'
 
 ## Serverless databases
